Library Management System User Interface
This folder contains the User interface of Library Management System . This is created by the Java-Fx UI framework. We use the Maven is chiefly used for Java-based projects, helping to download dependencies, which refers to the libraries or JAR files. The tool helps get the right JAR files for each project as there may be different versions of separate packages.